requirements.txt
# 核心依赖
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# AI依赖
pyyaml>=6.0.0
python-dotenv>=1.0.0

# 可选依赖（用于在线模式）
requests>=2.31.0

# 前端依赖（可选）
streamlit>=1.28.0

